Starting the Ford Dashboard.
This component is based around visualizing the vehical sales for
the Ford Motor Company sales of 2017-2018. There will be three
charts and three main button groups so the user can view different
aspects of the sales data.
The carts will include a bar chart, a line chart and a pack(bubble)
chart.