Added multiple elements (circles and paths) to show the relationship between years.
Also added a helper function called 'renderPath' that takes four
coordinates and returns a string for the d attribute of the path
element.

I also altered the calcExtentForAxis function so that when it calculates
the lowest and highest value from an array it will make the lower value
smaller and the higher value greater. This is so that when the values are
charted on the graphic they will not ride the top and bottom.