Added two methods to set state for the group of four buttons.
Before these four 'category' buttons would use setCategory to update
state when a click is detected but now the set state on two other
properties in state.
The two properties are tracking which item (or model of vehical) is
currently in 'focus' by the user. So one property tracks that models name
and the other contains the 'data' on that model. And that 'data' is
an object which holds all the specifics for that model.